Fix crash due to incorrect interpolation detection
Not all strings of the form ${...} are actually valid interpolations.
For instance the "=" character is not allowed currently, so that the
string "${foo:x=y}" must be considered a string *value* rather than an
interpolation.

The previous implementation crashed with for instance:

var1: y
var2: ${foo:x=${var1}}

This commit sets var2 to the string "${foo:x=y}".