Additional optimization to update() without force_add
This avoids invalidating the flags cache with `force_add=False`.

This gives a ~10% speed-up on the benchmark test_update_force_add that
updates the key ""a.a.a.a.a.a.a.a.a.a.a" (previous benchmark), and a
~50% speed-up on the new (and faster) benchmark that just updates "a".

This commit also fixes an issue with this benchmark, where re-using the
same config across tests was introducing dependencies between tests.