ensure hasattr returns False for invalid key types
Note we diverge __getattr__ (cfg.foo) and __getitem__ (cfg["foo"])
implementations here as the former attribute is implicitly always
a string, so we should not be validating types on it.

Note we don't overload the existing "validate_access" flag on
these methods, as we do want to validate config traversal
for the specified node, (for example checking for RecursionError).

Therefore we introduce a new "validate_key" flag to
select the separate behavior for __getattr__().

* omegaconf/dictconfig.py (__getattr__): Pass a new 'validate_key'
flag through __get_impl__ and __get_node__, setting it to false.
* tests/test_basic_ops_dict.py: Add a test case, checking
explicit hasattr() and implicit attribute.access syntax.

The significant tests for this change are:

  pytest tests/test_basic_ops_dict.py -k test_typed_hasattr
  pytest tests/test_basic_ops_dict.py -k test_getitem_with_invalid_key
  pytest tests/test_errors.py -k test_get_full_key_failure_in_format_and_raise